Move EP sharding into the sharding system
