+++ tensorrt_llm/_torch/auto_deploy/config/default.yaml
@@ -31,6 +31,7 @@ transforms:
     stage: pattern_matcher
   match_dense_moe_pattern:
     stage: pattern_matcher
+    run_shape_prop: true
   match_repeat_kv:
     stage: pattern_matcher
   match_eager_attention:
@@ -80,8 +81,6 @@ transforms:
   sharding_transform_executor:
     stage: sharding
     run_shape_prop: true
-  ep_sharding_mxfp4:
-    stage: sharding
   ############################################################################################
   # MOVE MODEL AND LOAD WEIGHTS
   ############################################################################################
