+++ tensorrt_llm/_torch/auto_deploy/utils/sharding_utils.py
@@ -652,6 +652,84 @@ def _insert_sharded_moe(
         dist_node.replace_input_with(dist_node, node)
 
 
+def _slice_expert_dim(gm: GraphModule, tensor_node: Node, lo: int, hi: int) -> Node:
+    """Return tensor_node[lo:hi, ...] via aten.slice along dim 0."""
+    lo = int(lo)
+    hi = int(hi)
+    with gm.graph.inserting_after(tensor_node):
+        # aten.slice.Tensor(self, dim, start, end, step)
+        return gm.graph.call_function(
+            torch.ops.aten.slice.Tensor,
+            args=(tensor_node, 0, lo, hi, 1),
+        )
+
+
+def _insert_sharded_mxfp4_mlp_ep(
+    gm: GraphModule,
+    node: Node,
+    rank: int,
+    world_size: int,
+):
+    """
+    Transform a call to auto_deploy::triton_mxfp4_mlp into:
+      - sharded expert parameters along dim 0 (this rank's slice),
+      - call to auto_deploy::triton_mxfp4_mlp_ep(..., local_lo, local_hi),
+      - followed by torch_dist_all_reduce.
+
+    Expects the original op signature:
+      (hidden_states,
+       router_weight, router_bias, top_k,
+       gate_up_blocks, gate_up_bias, gate_up_scales,
+       alpha, limit,
+       down_blocks, down_bias, down_scales)
+    """
+
+    IDX_GATE_UP_BLOCKS = 4
+    IDX_GATE_UP_BIAS = 5
+    IDX_GATE_UP_SCALES = 6
+    IDX_DOWN_BLOCKS = 9
+    IDX_DOWN_BIAS = 10
+    IDX_DOWN_SCALES = 11
+
+    gate_up_blocks_node = node.args[IDX_GATE_UP_BLOCKS]
+    num_experts = int(gate_up_blocks_node.meta["val"].shape[0])
+
+    # Compute per-rank [lower, upper)
+    base = num_experts // world_size
+    rem = num_experts % world_size
+    if rank < rem:
+        local_lo = rank * (base + 1)
+        local_hi = local_lo + (base + 1)
+    else:
+        local_lo = rem * (base + 1) + (rank - rem) * base
+        local_hi = local_lo + base
+
+    # Prepare new args with slices for this rank
+    args = list(node.args)
+    with gm.graph.inserting_before(node):
+        args[IDX_GATE_UP_BLOCKS] = _slice_expert_dim(
+            gm, args[IDX_GATE_UP_BLOCKS], local_lo, local_hi
+        )
+        args[IDX_GATE_UP_BIAS] = _slice_expert_dim(gm, args[IDX_GATE_UP_BIAS], local_lo, local_hi)
+        args[IDX_GATE_UP_SCALES] = _slice_expert_dim(
+            gm, args[IDX_GATE_UP_SCALES], local_lo, local_hi
+        )
+        args[IDX_DOWN_BLOCKS] = _slice_expert_dim(gm, args[IDX_DOWN_BLOCKS], local_lo, local_hi)
+        args[IDX_DOWN_BIAS] = _slice_expert_dim(gm, args[IDX_DOWN_BIAS], local_lo, local_hi)
+        args[IDX_DOWN_SCALES] = _slice_expert_dim(gm, args[IDX_DOWN_SCALES], local_lo, local_hi)
+
+    args_ep = tuple(args) + (int(world_size), int(rank))
+    node.target = torch.ops.auto_deploy.triton_mxfp4_mlp_ep.default
+    node.args = args_ep
+
+    # Add a dist all-reduce after the op (sum partial results across EP ranks)
+    with gm.graph.inserting_after(node):
+        red = gm.graph.call_function(torch.ops.auto_deploy.torch_dist_all_reduce, args=(node,))
+        node.replace_all_uses_with(red)
+        # keep dataflow: red(input=node)
+        red.replace_input_with(red, node)
+
+
 class EPShardingInfo(ShardingTransformInfo):
     """Configuration for EP sharding transformations."""
 
@@ -676,6 +754,20 @@ class EPShardingInfo(ShardingTransformInfo):
     def apply(self, gm: GraphModule, node: Node) -> None:
         """Apply EP sharding transformation to the graph module."""
         _insert_sharded_moe(gm, node, self.rank, self.world_size, [])
+
+
+class MXFP4EPShardingInfo(EPShardingInfo):
+    """GPT-OSS style MXFP4-specific EP sharding behavior."""
+
+    def validate(self, gm: GraphModule = None, node: Node = None) -> bool:
+        """Validate the transformation configuration."""
+        if not is_op(node, torch.ops.auto_deploy.triton_mxfp4_mlp):
+            ad_logger.warning(f"EP sharding is only supported for MOE nodes. Skipping {self}.")
+            return False
+        return True
+
+    def apply(self, gm: GraphModule, node: Node) -> None:
+        _insert_sharded_mxfp4_mlp_ep(gm, node, self.rank, self.world_size)
 
 
 class FP8EPShardingInfo(EPShardingInfo, QuantizationShardingMixin):
@@ -714,6 +806,7 @@ EP_SHARDING_RULES = [
     (lambda n: is_op(n, torch.ops.auto_deploy.torch_quant_fp8_moe), FP8EPShardingInfo),
     (lambda n: is_op(n, torch.ops.auto_deploy.torch_quant_nvfp4_moe), NVFP4EPShardingInfo),
     (lambda n: is_op(n, torch.ops.auto_deploy.torch_moe), EPShardingInfo),
+    (lambda n: is_op(n, torch.ops.auto_deploy.triton_mxfp4_mlp), MXFP4EPShardingInfo),
 ]
